requirements.txt
joblib
numpy
pandas
primer3-py
requests
scikit-learn
streamlit
